requirements.txt
absl-py==2.1.0
ale-py==0.8.1
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1660605382950/work
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
box2d-py==2.3.5
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1648854303569/work
certifi==2022.9.24
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1656783163977/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
cloudpickle==3.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
conda==22.9.0
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1649385138064/work
contourpy==1.3.0
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1657174084384/work
cycler==0.12.1
Cython==0.29.21
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1660619056300/work
decorator==4.4.2
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
etils==1.10.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1665301981797/work
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
glfw==2.7.0
google-pasta==0.2.0
grpcio==1.67.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.12.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1664214831245/work
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1662481590184/work
jax==0.4.35
jaxlib==0.4.35
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1659959867326/work
Jinja2==3.1.4
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1665540280048/work
jupyter_core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1658332693870/work
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
ml-dtypes==0.4.1
moviepy==1.0.3
mpmath==1.3.0
mujoco==3.2.4
namex==0.0.8
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.4.2
numpy==1.26.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==11.0.0
proglog==0.1.10
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1662384672173/work
protobuf==5.28.3
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1662356207792/work
psycopg2 @ file:///Users/runner/miniforge3/conda-bld/psycopg2-split_1640944574767/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1649385121707/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pygame==2.6.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1660666458521/work
PyOpenGL==3.1.7
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1648857351433/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2022.4
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1663830754951/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1656534056640/work
rich==13.9.3
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1653464701544/work
scipy==1.14.1
Shimmy==0.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stable_baselines3==2.3.2
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1664126450622/work
swig==4.2.1.post0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
torch==2.5.1
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1656938163103/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1663005918942/work
typing_extensions==4.12.2
tzdata==2024.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==3.1.0
wrapt==1.16.0
zipp==3.20.2